requirements.txt
matplotlib-inline==0.1.3
numpy==1.21.5
pandas==1.3.5
requests==2.27.1
streamlit==1.6.0